Add DeleteItem component
Created a DeleteItem component to handle the deletion of shop items. The
DeleteItem component uses a new deleteItem mutation to modify the
database and updates the Apollo client by retrieving all items using the
existing ALL_ITEMS query and filters out the deleted item. Updating the
Apollo client cache directly allows for the deletion to be reflected
instantaneously in the application, there is no subsequent database
query needed to display items post-deletion.