Add Pagination component
The Pagination component will be used to present the shop storefront
with a method of jumping through the various pages of items that are
going to be rendered on the main page. The genereated itemsConnection
prisma query is used to return the total number of items held in the
database.

The created Pagination Query uses the itemsConnection resolver to fetch
item count data. Total Item pages are calculated in the Pagination
component.