Add backend user signup functionality
The User graphql model was updated with data fields to store additional
information for user AUTH and token handling.

Cookies will be used to carry user login state in the User's session
across pages. Cookie-Parser express middle ware will be used to handle
this cookie creation.

A signup mutation was added to the graphql schema and a corresponding
Mutation was added to the application resolvers set. The signup resolver
handles the hashing of user password, user creation in database, and JWT
token creation.