Add Signin component
A signin Mutation and corresponding resolver was created to handle
retrieving user data from the database when attempting to sign in
using the Signing component. The Signin component uses the new Mutation
to retrieve user details and refetches the current user query in order
to render user details in the components that use that query.