Add backend password reset handling
Password request and reset mutations were created along with resolvers
for allowing a user to reqeust a reset token and to reset a password
using a valid reset token.

The request token is created using randomBytes package and gives a 1
hour expiry time limit.

The password reset mutation uses the generated Prisma updateUser
mutation to change the password data for the user and resets the token
values after usage. A new JWT is then generated with the new updated
user object.